Feat: 가족 및 가족 구성원 api 변경 (FCM 적용 보류)

+++ src/domain/family_member/family-member.service.ts
@@ -26,6 +26,10 @@ export class FamilyMemberService {
     createFamilyMemberDto: CreateFamilyMemberDto,
   ): Promise<number> {
     const user = await this.validateUser(userId);
+    // 이미 가족에 속해있는 경우 오류 발생
+    if (user.belongsToFamily) {
+      throw new FamilyMemberException(ResponseCode.FAMILY_MEMBER_ALREADY_EXIST);
+    }
     const family = await this.validateFamily(createFamilyMemberDto.familyId);
     const familyMember: FamilyMember = FamilyMember.createFamilyMember(
       createFamilyMemberDto.role,
@@ -33,6 +37,15 @@ export class FamilyMemberService {
       user,
     );
     const savedMember = await this.familyMemberRepository.save(familyMember);
+    family.memberNumber += 1;
+    // 가족에 속해 있다고 지정
+    await this.userRepository.update(user.id, {
+      belongsToFamily: () => 'BelongsToFamily + 1',
+    });
+    // 가족 구성원 수 증가
+    await this.familyRepository.update(family.id, {
+      memberNumber: () => 'Member_Number + 1',
+    });
     return savedMember.id;
   }
 
@@ -47,16 +60,33 @@ export class FamilyMemberService {
   }
 
   async deleteFamilyMember(familyMemberId: number) {
-    await this.validateFamilyMember(familyMemberId);
+    const familyMember = await this.familyMemberRepository.findOne({
+      where: { id: familyMemberId },
+      relations: ['family', 'user'],
+    });
+    if (!familyMember) {
+      throw new FamilyMemberException(ResponseCode.FAMILY_MEMBER_NOT_FOUND);
+    }
+    const family = await this.validateFamily(familyMember.family.id);
+    const user = await this.validateUser(familyMember.user.id);
+
     await this.familyMemberRepository.delete(familyMemberId);
+    await this.userRepository.update(user.id, {
+      belongsToFamily: () => 'BelongsToFamily - 1',
+    });
+    await this.familyRepository.update(family.id, {
+      memberNumber: () => 'Member_Number - 1',
+    });
   }
 
+  //유저 ID를 통한 가족 구성원 정보 반환
   async findFamilyMemberByUserId(
     userId: number,
   ): Promise<ResponseFamilyMemberDto> {
     await this.validateUser(userId);
     const familyMember = await this.familyMemberRepository.findOne({
       where: { user: { id: userId } },
+      relations: ['user'],
     });
     if (!familyMember) {
       throw new FamilyMemberException(ResponseCode.FAMILY_MEMBER_NOT_FOUND);
@@ -64,19 +94,43 @@ export class FamilyMemberService {
     return ResponseFamilyMemberDto.from(familyMember);
   }
 
+  //가족 구성원 ID를 통한 가족 구성원 정보 반환
+  async findFamilyMemberByMemberId(
+    familyMemberId: number,
+  ): Promise<ResponseFamilyMemberDto> {
+    const familyMember = await this.familyMemberRepository.findOne({
+      where: { id: familyMemberId },
+      relations: ['user'],
+    });
+    if (!familyMember) {
+      throw new FamilyMemberException(ResponseCode.FAMILY_MEMBER_NOT_FOUND);
+    }
+    return ResponseFamilyMemberDto.from(familyMember);
+  }
+
+  //가족 구성원 ID를 통한 가족 정보 반환
   async findFamilyByMemberId(
     familyMemberId: number,
   ): Promise<ResponseFamilyDto> {
-    const familyMember = await this.validateFamilyMember(familyMemberId);
+    const familyMember = await this.familyMemberRepository.findOne({
+      where: { id: familyMemberId },
+      relations: ['family'], // Family 테이블과 Join
+    });
+
+    if (!familyMember) {
+      throw new FamilyMemberException(ResponseCode.FAMILY_MEMBER_NOT_FOUND);
+    }
     return ResponseFamilyDto.from(familyMember.family);
   }
 
+  //가족 ID를 통한 모든 가족 구성원 정보 반환
   async findAllFamilyMemberByFamilyId(
     familyId: number,
   ): Promise<ResponseFamilyMemberDto[]> {
     await this.validateFamily(familyId);
     const familyMembers = await this.familyMemberRepository.find({
       where: { family: { id: familyId } },
+      relations: ['user'],
     });
     return familyMembers.map((familyMember) =>
       ResponseFamilyMemberDto.from(familyMember),
